Update patch set 1
Patch Set 1:

Hi Jim.

That is a good feature and these prints are very useful for a regular usage. But can we add some cli parameter to disable it for some cases? We use perf in our scripts and we grep perf output to get performance metrics. With this change we may receive a lot more output (e.g. when we run perf for several minutes) and we'll have to skip it.

Patch-set: 1
CC: Gerrit User 1016482 <1016482@d5d70762-12d0-45a1-890d-524b12d3f735>