Update patch set 4
Patch Set 4: Code-Review-1

Is this the fix for #1018, I think it's more likely a cleanup.
For timeout issue when removing the LUN, we should give an error response for all the outstanding primary tasks(submitted to backend or not).

Patch-set: 4
Reviewer: Gerrit User 1011204 <1011204@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1