lib/iscsi: Remove unnecesary adjustment of transfer length of subtask
As the previous patch shows, this adjustment of subtask->scsi.transfer_len
to subtask->scsi.length is not necessary.  It looks that this is
related with process_read_task_completion() which checks if
task->scsi.transfer_len is equal to task->scsi.length.  However,
for non-split case, there is no subtask and primary->scsi.transfer_len
is equal to task->scsi.length, and for split case, task is not equal
to primary.  Hence remove this adjustment.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: Id1602f69222f1946cd5bc5f24be76c74f197ae3d