lib/scsi: Remove adjustment of transfer length in spdk_scsi_task_process_null_lun()
As the previous patch shows, this adjustment of subtask->scsi.transfer_len
to subtask->scsi.length is not necessary.  It looks that this is
related with process_read_task_completion() which checks if
task->scsi.transfer_len is equal to task->scsi.length.  However,
for non-split case, there is no subtask and primary->scsi.transfer_len
is equal to task->scsi.length, and for split case, task is not equal
to primary.  Hence remove this adjustment.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I7355be9f595622df04781a9d49c95777eb804643