lib/iscsi: Reduce timeout to wait logout from initiator to 5 seconds
After this patch series, iSCSI hotplug works as expected but connecions
are stil active.  This is desirable behavior.  But for some case, even
if connections are active, initiator doesn't reply to asynchronous
logout request.  30 seconds timeout is too long for this case.

This patch reduces the timeout from 30 seconds to 5 seconds not to
process this timeout as error.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I33c56ca6d383c0272bbb78e8876b548fa4eae503