Update patch set 4
Patch Set 4: Code-Review-1

Why did the test not fail in this case?

I suspect this may be happening in a lot of tests.  We need to fix the test script as well so that the test fails if it hits a null pointer dereference.

Patch-set: 4
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1