Update patch set 21
Patch Set 21: Code-Review-1

Hi Hailiang,

Thanks for your work on this! I see that you have done a really good job following the test and application structure for the fuzzers.

iscsi is a slightly different situation than the other transports. I don't think that we will be able to use the common APIs like we did with those ones. The iscsi stuff is a lot more restrictive than NVMe-oF and vhost. For example, when you submit an invalid opcode with the iscsi_pdu_payload_handle function and it has a bad opcode, it never makes it over to the target application. It gets failed before being submitted. The same is true for a lot of error conditions checked by the iscsi_pdu_payload_op_* family of functions. Is it possible to change this to use some of the lower level APIs like spdk_iscsi_conn_free_pdu?

Another issue is the fact that we are using iscsiadm as the true initiator. There are probably a lot of issues related to iscsiadm failing to send malformed requests that will prevent us from properly fuzzing the target. Ultimately, I think the goal here is to have a completely custom initiator. But that is obviously a lot of work.

Could you maybe start by modifying this application to not use the iscsi_pdu_payload_handle function and try to use lower level APIs to communicate to the target while still relying on iscsiadm?

From there, we can look at how many requests are actually making it to the target side and evaluate how to replace the portion that iscsiadm is controlling in the future.

Patch-set: 21
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1