nvme/transport: don't return from NVME_TRANSPORT_CALL
Returning directly from the NVME_TRANSPORT_CALL macro is really slick
and allows us to avoid a lot of lines of code, but it doesn't allow us
to do anything within the transport functions after calling the macro.

There are several places where we want to be able to modify state within
these functions before returning.

Change-Id: Id4183ade3259e6c259218bef637ab9b94084cce5
Signed-off-by: Seth Howell <seth.howell@intel.com>