app/rpc: Add thread_get_threads RPC to retrieve threads' core affinity
Thread's core affinity had not been visible to users. This patch
adds thread_get_threads RPC to retrieve thread's core affinity.

Further information may be added to the RPC and so make the name of
the RPC general.

We can share context structure between thread_get_stats and
thread_get_threads RPCs and do it in this patch by renaming
the name of the structure by rpc_get_threads_ctx.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I634da46c411d648837538ebf227074d307273c97