test/ocf: don't kill bdevperf in stats.sh
There is no need to run bdevperf for two minutes then try to kill it in
this test. There were intermittent failures on the test pool where we
were unable to use sigterm to kill bdevperf after this test causing the
test suite to hang.
I'm just removing the killprocess call entirely and decreasing the
amount of time we run bdevperf.

Change-Id: Ie8bd9252fabbce55f921bdb7db1e44690f8bf93e
Signed-off-by: Seth Howell <seth.howell@intel.com>