Update patch set 29
Patch Set 29:

> 
> Hi Jim,
> 
> Can you specify which code is complicated, I think that the current code is not easy to maintain. And this one could be much easier for the further maintain.

This patch could be much easier, but we don't know if it breaks anything.  We have no unit tests for much of this code.  The current code is very complicated, which is why it is not easy to maintain.  If we have good unit tests, then we can refactor the code and be confident that it does not break any existing functionality.

Patch-set: 29
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>