Update patch set 29
Patch Set 29:

> Patch Set 29:
> 
> > 
> > Hi Jim,
> > 
> > Can you specify which code is complicated, I think that the current code is not easy to maintain. And this one could be much easier for the further maintain.
> 
> This patch could be much easier, but we don't know if it breaks anything.  We have no unit tests for much of this code.  The current code is very complicated, which is why it is not easy to maintain.  If we have good unit tests, then we can refactor the code and be confident that it does not break any existing functionality.

Hi Jim,

The change affects the original ut, and the ut is updated to make sure it works, so it will not break anything. 

Without this patch change, our code in the master branch will become harder to maintain in the future. You can see data_in_cnt will be spread many spaces. Generally, the idea is clear. We do not need to track the primary read task for the data_in_cnt. If the large read I/O needs split, the primary task need not be put into scsi layer to execute. And without it, our code will always need to check whether the primary task needs to be put or not. This makes the logic very worse though the logic previously is designed and implemented by me. But now I have the good implementation to fix this and make it easier.

Patch-set: 29
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>