[RFC] SPDK : Preallocate 1GB mapping page for some of the indices.
For TCMalloc regions which we register with spdk at runtime in the MMapHook, we
need to ensure that SPDK doesn't do any allocations in that path otherwise we
will hit a livelock situation. MmapHook is invoked when TCMalloc is out of free
memory and needs to get more memory from the system, for the hugepage case it
gets via mmap.

With this change I am adding support for preallocation of these 1GB mapping
pages for virtual region as specified by the user. The idea being we will
allocate the 1GB mapping page for the whole address space that TCmalloc will
use.
Two parameters need to be specified,
  start_addr - starting virtual address of the region.
  num_entries - number of indices that need to be populated. Note that each
entry denotes a contiguous 1GB region.

So for a 16GB region we will specify num_entries as 16. Note that we allocate a
4K block for each entry, so for a 16G hugepage configuration this will
translate into 64K of memory being preallocated.

Note that these parameters are hardcoded in the current change, but my plan is
to expose this via spdk_env_opts, so will add couple of fields in that struct,
that the application can set before calling spdk_env_init.

Change-Id: I92c62b17361a29cc33fb9522f737b626a950afad
Signed-off-by: Alok N Kataria <alok.kataria@nutanix.com>