Update patch set 31
Patch Set 31:

(1 comment)

Hi Ziye,

I agree that this new code is simpler.  But this is some very complicated code that you are changing in this patch.  We have had lots of intermittent failures in iSCSI.  So I would like to see us have more and better unit tests in place, before we change all of this code.  I pointed out a couple of areas of the old code that weren't getting hit by unit tests at all.  Right now, the patch is passing the test pool, but I'm not confident that there couldn't be bugs that won't hit until it gets merged and we see intermittent failures in the test pool.

I'm not saying that your patch is wrong.  I like the idea of the patch and want to make the code simpler.  I'm just asking that you take the lead here on improving the unit tests so we can be more confident that changes to this code are solid.

Thanks,

-Jim

Patch-set: 31
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>