[RFC] bdev: add spdk_bdev_comparev_and_writev_blocks
We will only supported a vectored variant of
compare-and-write for now.

This still needs to be broken up into some smaller
patches - especially some of the bdev_ut refactoring.

This does no locking for now.  I think I'll add the
locking in a separate patch after this one.  This
patch by itself should be enough to enable NVMe-oF
and NVMe driver development.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I5bd075c912de60090e19cf8fced19c4879fcc900