vhost-blk: add resize bdev support
This will allow us to resize the backend bdev of vhost-blk and notify the guest
OS that the capactiy of virtio-blk disk has been resized.

The spdk api entry is `spdk_bdev_notify_blockcnt_change`.
Any bdev if used as vhost-blk backend may need to implement a rpc that
calls this function.

This patch works with the dpdk patch:
https://patches.dpdk.org/patch/63500/

Signed-off-by: Li Feng <fengli@smartx.com>
Change-Id: Iaf59dc3da09c9d64c5a9d7ee3d988025ec70307a
Signed-off-by: Li Feng <fengli@smartx.com>