lib/iscsi: Do not remove pending or being flushed PDU/tasks for LUN hotplug
We ensure current pending tasks are aborted and no new task or PDU
is pending after LUN is unplugged now.  Then we have to ensure all
task completions and aborts are sent to initiator.

However, we had removed pending tasks without any notification to
initiator.  For this implementation, we are not surprised if initiator
waits for in-flight commands for a long time.

Due to recent refinements of LUN hotplug, we can wait safely that
all existing tasks complete or abort.

Hence this patch removes _iscsi_conn_free_tasks() from
_iscsi_conn_remove_lun(), and then simplify and then inline
_iscsi_conn_free_tasks() into iscsi_conn_free_tasks(), and
updates unit tests accordingly.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I869b79a7c93d2e8a4a1577cc20d3b466548dfaaa