lib/blob: separate cluster allocation from insertion
Cluster allocation is two step process:
1) a bit in used_clusters is set
2) corresponding LBA is written into that cluster

1) and 2) were done using _spdk_bs_allocate_cluster(),
if last argument 'update_map' was set.
2) was done specifically by _spdk_blob_insert_cluster().

Cluster allocation in blobstore occurs under following cases:
a) thick provisioned blob is resized
b) first write to a cluster occurs on thin provisioned blob

In case a), both 1) and 2) were performed since we already were
on md_thread and results could be immidiate.

In case b), 1) occured immidietly on thread the io was issued.
When the thin provisioned blob was backed by another blob
(for example io to a clone), first read of that cluster
is performed from backing blob.
Only after this is complete, step 2) occurs on md_thread.

Current implementation of 1) and 2) is synchronous.
In the future 2) will be asynchronous to enable
updates of just part of blob metadata.

This patch parepares for that change, by separating 1)
and 2). Where 2) will be made async.

No change to blob behaviour should occur with this patch.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I615d013b77607e9e7e5060ec804cdddb7ad0e6f1