lib/blob: make _spdk_blob_resize asynchronous
Insertion of the cluster allocation will be writting
additional metadata, thus be asynchronous.

Due to above, all paths that cause cluster allocation
need to be made asynchronous.
When creating a blob, resizing a blob or deleting a blob,
internal _spdk_blob_resize() occurs, followed by already
asynchronous _spdk_blob_persist().

This patch changes _spdk_blob_resize() to be asynchronous.
Similar to _spdk_blob_persist() it now accepts the same,
spdk_bs_sequence in creation or deletion of a blob.

Resizing a blob didn't do blob persist, so had no need for
sequence. This patch adds it, in this path too.

In future this sequence will be used to issue writes
to additional metadata.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I080d8780a5fb956cc4be62b2c7765d6ee4a7e183