lib/blob: insert cluster now accepts batch as parameter
Inserting a cluster will in the future write additional
metadata to device.

In preparation for this, _spdk_blob_insert_cluster()
will now accept spdk_bs_batch. Nothing is done on it yet.

This requires for _spdk_blob_resize() to change its path
when blob is thin provisioned to await completion of
all cluster inserts.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I0ce84cd894610c21be74554ff8961479046d3dba