Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Gang,
> 
> Are you sure this patch can fix the issue? 
> 
> Thanks
> 
> Liang

Test method below:

1. One application (rpc_client.sh) to hold the sock and lock file
2. Another instance of rpc_client.sh without my change, it will hang at waiting the Semaphore
2*. With fix, another instance of rpc_client will exit in the case of not able to getting the Sock (already in use)

This is to fix the Time out due to the infinitely wait of Semaphore which is never got increased in the error case.

For a follow up issue like the test case itself is failed due to in use sock and lock file, it's another one. This patch is to fix the time out.

Thanks,
Gang

Patch-set: 1