scripts/vagrant: do not treat SPDK_VAGRANT_DISTRO as array
Dash and double dash ("-" and "--") getopts params are safely
shifted using "shift" and "OPTIND". After that it's safe to
concatenate param array to string using "$*", especially that
there is only 1 positional parameter, which is distro version.

Previous approach ("$@" as array) made "export" impossible and
as a result VM was created with Fedora 28 distro on each run.

Change-Id: I48fa588220f8b7f46c561d480ee32920847c7903
Signed-off-by: Karol Latecki <karol.latecki@intel.com>