Update patch set 2
Patch Set 2: Code-Review-1

Agreed with Karol.
This shellcheck seems more like a warning, since the behavior of passing scripts args to particular function is not desired in majority of cases. Disabling this check might be better, than reminding in most cases and requiring to 'work around' it.

This script could be improved to always require the signal type to be used or use functions available in common scripts (autotest_common.sh or nvmf/common.sh), but this is not required for this patch.

Patch-set: 2
Reviewer: Gerrit User 1011221 <1011221@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1