Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> Changing the name of "delay_pcie_doorbell" breaks the NVMe driver API.  Anyone who has started setting this flag (since it's off by default) would be broken once this patch goes in.
> 
> What if we add something like:
> 
> #define delay_pcie_doorbell delay_cmd_submit
> 
> I think this #define would be rather harmless.  An unnamed union for the two names would work too and avoid any possible name collisions, but someone using the name "delay_pcie_doorbell" anywhere else seems pretty remote to me.

Hi Jim,

We had a discussion on this with Ben in the last patch in series: https://review.gerrithub.io/c/spdk/spdk/+/462585
He was OK with renaming. ABI is kept and there is a small chance that many apps use it.
But I also like the approach with unnamed union. It seems safe and simple. I will update commit.

BR, Evgeniy.

Patch-set: 3