Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Hi Gang,
> 
> I read my comment again.
> 
> My comment is one idea and I'm afraid that I miss critical something.
> So please take it as just an idea.
> 
> -Shuhei

Also copied here:

This _available() check is different which is to confirm the NIC device is there by the index.

The newly added is_ip_address_in_use() is to check whether the IP is in-use on an existing NIC (which is always available there).

These two functions are different. Are you talking about this in your latest comment?

Thanks, Gang

Patch-set: 6